Refactor: move code to separate module
